requirements.txt
pymongo
flask
uwsgi
markdown
python-dotenv
